requirements.txt
flask
flask-sqlalchemy
sqlalchemy
python-dotenv
mysql-connector-python
pytest
pytest-flask
